requirements.txt
scipion-pyworkflow
scipion-em
scipion-app